refactor: follow entity 위치 이동
