+++ src/follows/entities/follow.entity.ts
@@ -1,1 +1,25 @@
-export class Follow {}
+import { User } from '@src/users/entities/user.entity';
+import {
+  CreateDateColumn,
+  Entity,
+  JoinColumn,
+  ManyToOne,
+  PrimaryGeneratedColumn,
+} from 'typeorm';
+
+@Entity('follows')
+export class Follow {
+  @PrimaryGeneratedColumn()
+  id: number;
+
+  @ManyToOne(() => User, (user) => user.followings)
+  @JoinColumn({ name: 'following_user_id' })
+  follower: User;
+
+  @ManyToOne(() => User, (user) => user.followers)
+  @JoinColumn({ name: 'followed_user_id' })
+  following: User;
+
+  @CreateDateColumn()
+  created_at: Date;
+}
